+++ src/utils/index.ts
@@ -1,0 +1,10 @@
+// src/utils/index.ts
+
+// Валидация данных
+export { validateUniqueIds } from "./validate-data";
+export * from "./dom";
+export { initTabs } from "@utils/tv01";
+
+// Будущие утилиты (пример)
+// export { formatDate } from './formatDate';
+// export { slugify } from './slugify';